backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub==0.36.2",
    "ipykernel>=7.2.0",
    "langgraph>=1.1.1",
    "moshi>=0.2.13",
    "python-dotenv>=1.2.2",
    "sounddevice>=0.5.0",
    "tavily-python>=0.7.23",
]
